Не работает отправление поста. Округление координат работает правильно. Начата работа по загрузке списка постов. Начата работа над адаптером

+++ app/src/main/java/ru/maxpek/friendslinkup/entity/PostRemoteKeyEntity.kt
@@ -4,10 +4,10 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity
-data class PostRemoteKeyEntity (
+data class PostRemoteKeyEntity(
     @PrimaryKey
     val type: KeyType,
-    val id: Long,
+    val id: Int,
 ) {
     enum class KeyType {
         AFTER, BEFORE
